Layout Functions Moved To DetailViewModel
